Simulation::Tick is seriously broken so that mesh is never polled but might move to another platform tomorrow. Got Win32 compiling - mostly missing pch inclusions in new .cpps. Seriously considering making all time variables doubles.
git-svn-id: https://crag.googlecode.com/svn/trunk@23 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/app/App.h
@@ -41,5 +41,5 @@ namespace app
 	void Sleep(TimeType t = 0);
 
 	double TimeTypeToSeconds(TimeType t);
-	TimeType SecondsToTimeType(float s);
+	TimeType SecondsToTimeType(double s);
 }
